migrate detail form section to separate component
